refactor(e2e): move rule cleanup into a fixture, tighten the sampling-rate timeouts
Cleanup moves out of the spec's try/finally into an opt-in
`automationRulesCleanup` fixture. Playwright runs fixture teardown regardless
of outcome, so it cannot be skipped by a failure; it receives a settled
testInfo, so shouldLeaveArtifacts works there (inside a test's own `finally`
the status is not final and still reads 'passed' on a failing test); and it
cannot swallow the test's error the way a `return` in `finally` does. All three
were bugs this spec hit in review. Keeps the exclusive-ownership filter, since
rule deletion is workspace-global.

Timeouts sized to the work rather than padded. The inner waits run concurrently
across the batch, so the worst case is one trace's chain, not the sum over 30:
test 900s -> 300s, control poll 240s -> 120s, settle 120s -> 90s. Observed
runtime is 29-46s. The previous 900s ceiling mostly meant a stalled run sat
there for 15 minutes before reporting.
