+++ tests_end_to_end/e2e/fixtures/index.ts
@@ -1,4 +1,4 @@
-export { test, expect } from './token-usage-spans.fixture';
+export { test, expect } from './automation-rules.fixture';
 export type { ProjectFixtures } from './project.fixture';
 export type { ScratchDir, ScratchDirFixtures } from './scratch-dir.fixture';
 export type {
@@ -83,4 +83,5 @@ export type {
   TokenUsageSpansRef,
   TokenUsageSpansFixtures,
 } from './token-usage-spans.fixture';
+export type { AutomationRulesCleanupFixtures } from './automation-rules.fixture';
 export type { ProjectRef } from '../core/backend';
